feat(wrpc) websocket server transmit queue/thread
server WebSockets are actually the request's socket, so they can't be
written to from the wrong context (timer, other requests).  A thread
spawned from the request-handling function is ok.

Add a semaphore-based queue to allow other contexts to send data to a
transmit thread.  Since this is only for the actual socket operation,
everything before that (fetch service/rpc definition, encode args,
annotate the packet, etc) happens in the "sending" context, improving
resource usage.  So far this is only within the same nginx worker.